requirements.txt
boto3==1.26.59
botocore==1.29.59
cachelib==0.9.0
certifi==2022.12.7
charset-normalizer==3.0.1
chartjs==1.2
click==8.1.3
Flask==2.2.2
Flask-Caching==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
praw==7.6.1
prawcore==2.3.0
python-dateutil==2.8.2
python-decouple==3.7
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
update-checker==0.18.0
urllib3==1.26.14
websocket-client==1.5.1
Werkzeug==2.2.2
